Ass 2/IR_ASSIGNMENT_2/requirements.txt
absl-py==2.1.0
anyio==3.7.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
astunparse==1.6.3
async-lru==2.0.2
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
bs4==0.0.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.7
colorama==0.4.6
comm==0.1.3
contourpy==1.1.0
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
exceptiongroup==1.1.1
executing==1.2.0
fastjsonschema==2.17.1
flatbuffers==24.3.25
fonttools==4.40.0
fqdn==1.5.1
gast==0.5.4
google-pasta==0.2.0
grpcio==1.62.2
h5py==3.11.0
idna==3.4
ipykernel==6.23.2
ipython==8.14.0
ipython-genutils==0.2.0
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter-lsp==2.2.0
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyter_server==2.6.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.2
jupyterlab-pygments==0.2.2
jupyterlab_server==2.23.0
keras==3.2.1
kiwisolver==1.4.4
Levenshtein==0.25.1
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matlab==0.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
memory-profiler==0.61.0
mistune==2.0.5
ml-dtypes==0.3.2
namex==0.0.8
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.5.0
nbformat==5.9.0
nest-asyncio==1.5.6
nltk==3.8.1
notebook==6.5.4
notebook_shim==0.2.3
numpy==1.24.3
opt-einsum==3.3.0
optree==0.11.0
overrides==7.3.1
packaging==23.1
pandas==2.2.2
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.5.3
prometheus-client==0.17.0
prompt-toolkit==3.0.38
protobuf==4.25.3
psutil==5.9.5
pure-eval==0.2.2
pyahocorasick==2.1.0
pycparser==2.21
Pygments==2.15.1
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-json-logger==2.0.7
python-Levenshtein==0.25.1
pytz==2024.1
pywin32==306
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.1.0
rapidfuzz==3.8.1
regex==2023.12.25
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
scikit-learn==1.4.2
scipy==1.13.0
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
stack-data==0.6.2
tabulate==0.9.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-intel==2.16.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
terminado==0.17.1
threadpoolctl==3.4.0
timeout-decorator==0.5.0
tinycss2==1.2.1
tomli==2.0.1
tornado==6.3.2
tqdm==4.66.2
traitlets==5.9.0
typing_extensions==4.6.3
tzdata==2024.1
uri-template==1.2.0
urllib3==2.0.3
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.3
Werkzeug==3.0.2
wrapt==1.16.0
